Resolves #84 created login page as the index.js
